complete

+++ src/main/java/lang/print/gaps/task5/NumberDivider.java
@@ -3,7 +3,7 @@ package lang.print.gaps.task5;
 public class NumberDivider {
     void divide(int toBeDivided) {
         float result;
-        result = toBeDivided / 5;
+        result = (float) toBeDivided / 5;
         System.out.println(result);
     }
 }
